# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERP V2 API",
        "description": "Comprehensive ERP System API Documentation with Multi-tenancy and Role-based Access Control",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.example.com/api/v1",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/admin/plans": {
            "get": {
                "tags": [
                    "Admin/Plans"
                ],
                "summary": "Get all plans",
                "operationId": "4b58fcd6d13184c5bf03c38a882f8a8e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin/Plans"
                ],
                "summary": "Create a new plan",
                "operationId": "74e431da7a2c9aea218a525ed2bff36a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePlanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/plans/{id}/change-status": {
            "patch": {
                "tags": [
                    "Admin/Plans"
                ],
                "summary": "Toggle plan active status",
                "operationId": "19d30f7ec0c5a558c41861c08d7ffd98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePlanRequest": {
                "required": [
                    "name",
                    "price",
                    "billing_cycle"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Premium Plan"
                    },
                    "description": {
                        "type": "string",
                        "example": "Full access to all features",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "billing_cycle": {
                        "type": "string",
                        "example": "monthly",
                        "enum": [
                            "monthly",
                            "yearly",
                            "lifetime"
                        ]
                    },
                    "max_users": {
                        "type": "integer",
                        "example": 50,
                        "nullable": true
                    },
                    "max_branches": {
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Advanced Reporting",
                            "API Access"
                        ],
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_popular": {
                        "type": "boolean",
                        "example": false
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {your_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}